Created navmesh for enemy AI
Something to keep in mind is that it may not be perfect and only accounts for walkable area (based on the selected criteria when I baked). Meaning jumpable areas are not considered and the enemy ai will not try to jump onto it. The navmesh can be found in Asylum game object